Persist output names in cell attributes
This is needed to be able to allow for rendering cells without starting
the engine.
